# Improve htable_bench output